Fixing performance-enum-size

+++ src/lib/kosinski.cc
@@ -46,7 +46,7 @@ class kosinski_internal {
         using descriptor_t        = uint16_t;
         using descriptor_endian_t = little_endian;
         using sliding_window_t    = sliding_window<kosinski_adaptor>;
-        enum class edge_type : size_t {
+        enum class edge_type : uint8_t {
             invalid,
             terminator,
             symbolwise,
