+++ src/lib/lzkn1.cc
@@ -47,7 +47,7 @@ class lzkn1_internal {
         using descriptor_t        = uint8_t;
         using descriptor_endian_t = big_endian;
         using sliding_window_t    = sliding_window<lzkn1_adaptor>;
-        enum class edge_type : size_t {
+        enum class edge_type : uint8_t {
             invalid,
             terminator,
             symbolwise,
